autoSAWP2

+++ include/driver.h
@@ -5,7 +5,7 @@ namespace driverControl {
 
 	namespace driveCurve {
 		constexpr int DEAD_ZONE = 2;
-		constexpr int DRIVE_CURVE[] = {0, 0.55, 0.45, 1};
+		constexpr double DRIVE_CURVE[] = {0, 0.55, 0.45, 1};
 		constexpr int MIN_OUTPUT = 2;
 		constexpr int SCALE_INPUT = 127;
 		constexpr int SCALE_OUTPUT = 127;
